Fix full load hours
I couldn't reproduce the full load hours previously set (1956.164053)
even with older versions of python (checked down to 3.8) and pandas
(checked down to 1.3.5), so I assume, that the value was incorrect. It
was possibly not checked, as the example tests are not in the tests
directory but the example directory.